Restructure GDS I/O into dedicated io/ module with read/write submodules
Consolidates scattered I/O code into a cleaner module hierarchy:
- Merge writer.rs and write portions of utils/io.rs into io/write/
- Move read code (from_gds, RecordReader) into io/read/
- Move gds_format.rs and svg.rs into io/write/
- Remove now-empty utils/ directory
- Add GDSDataType::byte_size/record_size, record_header helper, and
  write_i32_record to reduce duplication in serialization code
- Add default implementations to GdsWriter trait methods
- Rename combine_record_and_data_type to record_head
