+++ crates/gdsr/src/io/read/mod.rs
@@ -1,269 +1,14 @@
-use std::convert::TryFrom;
 use std::fs::File;
 use std::io::{self, BufReader, Read};
 
-use chrono::{Datelike, Local, Timelike};
-
 use crate::cell::Cell;
-use crate::config::gds_file_types::{
-    GDSDataType, GDSRecord, GDSRecordData, combine_record_and_data_type,
-};
+use crate::config::gds_file_types::{GDSDataType, GDSRecord, GDSRecordData};
 use crate::elements::text::get_presentations_from_value;
 use crate::elements::{GdsBox, Node, Path, PathType, Polygon, Reference, Text};
 use crate::error::GdsError;
 use crate::geometry::round_to_decimals;
 use crate::library::Library;
-use crate::utils::gds_format::{eight_byte_real, write_u16_array_as_big_endian};
 use crate::{DEFAULT_INTEGER_UNITS, DataType, Instance, Layer, Point, Unit};
-
-pub fn write_gds_head_to_file(
-    library_name: &str,
-    user_units: f64,
-    db_units: f64,
-    buffer: &mut impl std::io::Write,
-) -> Result<(), GdsError> {
-    let now = Local::now();
-    let timestamp = now.naive_utc();
-
-    let head_start = [
-        6,
-        combine_record_and_data_type(GDSRecord::Header, GDSDataType::TwoByteSignedInteger),
-        0x0258,
-        28,
-        combine_record_and_data_type(GDSRecord::BgnLib, GDSDataType::TwoByteSignedInteger),
-        timestamp.year() as u16,
-        timestamp.month() as u16,
-        timestamp.day() as u16,
-        timestamp.hour() as u16,
-        timestamp.minute() as u16,
-        timestamp.second() as u16,
-        timestamp.year() as u16,
-        timestamp.month() as u16,
-        timestamp.day() as u16,
-        timestamp.hour() as u16,
-        timestamp.minute() as u16,
-        timestamp.second() as u16,
-    ];
-
-    write_u16_array_to_file(buffer, &head_start)?;
-
-    write_string_with_record_to_file(buffer, GDSRecord::LibName, library_name)?;
-
-    let head_units = [
-        20,
-        combine_record_and_data_type(GDSRecord::Units, GDSDataType::EightByteReal),
-    ];
-    write_u16_array_to_file(buffer, &head_units)?;
-
-    write_float_to_eight_byte_real_to_file(buffer, user_units)?;
-    write_float_to_eight_byte_real_to_file(buffer, db_units)
-}
-
-pub fn write_gds_tail_to_file(buffer: &mut impl std::io::Write) -> Result<(), GdsError> {
-    let tail = [
-        4,
-        combine_record_and_data_type(GDSRecord::EndLib, GDSDataType::NoData),
-    ];
-    write_u16_array_to_file(buffer, &tail)
-}
-
-pub fn write_u16_array_to_file(
-    buffer: &mut impl std::io::Write,
-    array: &[u16],
-) -> Result<(), GdsError> {
-    Ok(write_u16_array_as_big_endian(buffer, array)?)
-}
-
-pub fn write_float_to_eight_byte_real_to_file(
-    buffer: &mut impl std::io::Write,
-    value: f64,
-) -> Result<(), GdsError> {
-    let value = eight_byte_real(value);
-    Ok(buffer.write_all(&value)?)
-}
-
-pub const MAX_POINTS: usize = 8191;
-pub const MAX_LAYER: u16 = 255;
-pub const MAX_DATA_TYPE: u16 = 255;
-pub const MAX_STRING_LENGTH: usize = 512;
-pub const MAX_STRUCTURE_NAME_LENGTH: usize = 32;
-pub const MAX_COL_ROW: u32 = 32767;
-pub const MIN_POLYGON_POINTS: usize = 4;
-
-/// Returns an `InvalidInput` error if the layer is out of the GDS2 spec range (0-255).
-pub fn validate_layer(layer: Layer) -> Result<(), GdsError> {
-    if layer.value() > MAX_LAYER {
-        return Err(GdsError::ValidationError {
-            message: format!(
-                "Layer {} exceeds maximum value of {MAX_LAYER}",
-                layer.value()
-            ),
-        });
-    }
-    Ok(())
-}
-
-/// Returns a validation error if the data type is out of the GDS2 spec range (0-255).
-pub fn validate_data_type(data_type: DataType) -> Result<(), GdsError> {
-    if data_type.value() > MAX_DATA_TYPE {
-        return Err(GdsError::ValidationError {
-            message: format!(
-                "Data type {} exceeds maximum value of {MAX_DATA_TYPE}",
-                data_type.value()
-            ),
-        });
-    }
-    Ok(())
-}
-
-/// Returns a validation error if the string exceeds 512 characters.
-pub fn validate_string_length(s: &str) -> Result<(), GdsError> {
-    if s.len() > MAX_STRING_LENGTH {
-        return Err(GdsError::ValidationError {
-            message: format!(
-                "String length {} exceeds maximum of {MAX_STRING_LENGTH} characters",
-                s.len()
-            ),
-        });
-    }
-    Ok(())
-}
-
-/// Returns a validation error if the structure name exceeds 32 characters or contains
-/// invalid characters (only alphanumeric, `_`, `?`, `$` are allowed).
-pub fn validate_structure_name(name: &str) -> Result<(), GdsError> {
-    if name.len() > MAX_STRUCTURE_NAME_LENGTH {
-        return Err(GdsError::ValidationError {
-            message: format!(
-                "Structure name length {} exceeds maximum of {MAX_STRUCTURE_NAME_LENGTH} characters",
-                name.len()
-            ),
-        });
-    }
-    if let Some(c) = name
-        .chars()
-        .find(|c| !c.is_ascii_alphanumeric() && *c != '_' && *c != '?' && *c != '$')
-    {
-        return Err(GdsError::ValidationError {
-            message: format!("Structure name contains invalid character: '{c}'"),
-        });
-    }
-    Ok(())
-}
-
-/// Returns a validation error if columns or rows exceed 32767.
-pub fn validate_col_row(columns: u32, rows: u32) -> Result<(), GdsError> {
-    if columns > MAX_COL_ROW {
-        return Err(GdsError::ValidationError {
-            message: format!("Column count {columns} exceeds maximum value of {MAX_COL_ROW}"),
-        });
-    }
-    if rows > MAX_COL_ROW {
-        return Err(GdsError::ValidationError {
-            message: format!("Row count {rows} exceeds maximum value of {MAX_COL_ROW}"),
-        });
-    }
-    Ok(())
-}
-
-pub fn write_points_to_file(
-    buffer: &mut impl std::io::Write,
-    points: &[Point],
-    database_units: f64,
-) -> Result<(), GdsError> {
-    let num_points = points.len().min(MAX_POINTS);
-
-    let record_size = 4 + (num_points * 8) as u16;
-    let xy_header_buffer = [
-        record_size,
-        combine_record_and_data_type(GDSRecord::XY, GDSDataType::FourByteSignedInteger),
-    ];
-
-    write_u16_array_to_file(buffer, &xy_header_buffer)?;
-
-    for point in points.iter().take(num_points) {
-        let point = point.to_integer_unit();
-        let x_real = point.x().absolute_value();
-        let y_real = point.y().absolute_value();
-
-        let scaled_x = (x_real / database_units).round() as i32;
-        let scaled_y = (y_real / database_units).round() as i32;
-
-        buffer.write_all(&scaled_x.to_be_bytes())?;
-        buffer.write_all(&scaled_y.to_be_bytes())?;
-    }
-
-    Ok(())
-}
-
-pub fn write_element_tail_to_file(buffer: &mut impl std::io::Write) -> Result<(), GdsError> {
-    let tail = [
-        4,
-        combine_record_and_data_type(GDSRecord::EndEl, GDSDataType::NoData),
-    ];
-    write_u16_array_to_file(buffer, &tail)
-}
-
-pub fn write_string_with_record_to_file(
-    buffer: &mut impl std::io::Write,
-    record: GDSRecord,
-    string: &str,
-) -> Result<(), GdsError> {
-    let byte_len = string.len();
-    let padded_len = byte_len + (byte_len % 2);
-
-    let string_start = [
-        (4 + padded_len) as u16,
-        combine_record_and_data_type(record, GDSDataType::AsciiString),
-    ];
-
-    write_u16_array_to_file(buffer, &string_start)?;
-
-    buffer.write_all(string.as_bytes())?;
-    if byte_len % 2 != 0 {
-        buffer.write_all(&[0])?;
-    }
-
-    Ok(())
-}
-
-pub fn write_transformation_to_file(
-    buffer: &mut impl std::io::Write,
-    angle: f64,
-    magnification: f64,
-    x_reflection: bool,
-) -> Result<(), GdsError> {
-    let transform_applied = angle != 0.0 || magnification != 1.0 || x_reflection;
-    if transform_applied {
-        let buffer_flags = [
-            6,
-            combine_record_and_data_type(GDSRecord::STrans, GDSDataType::BitArray),
-            if x_reflection { 0x8000 } else { 0x0000 },
-        ];
-
-        write_u16_array_to_file(buffer, &buffer_flags)?;
-
-        if magnification != 1.0 {
-            let buffer_mag = [
-                12,
-                combine_record_and_data_type(GDSRecord::Mag, GDSDataType::EightByteReal),
-            ];
-            write_u16_array_to_file(buffer, &buffer_mag)?;
-            write_float_to_eight_byte_real_to_file(buffer, magnification)?;
-        }
-
-        if angle != 0.0 {
-            let buffer_rot = [
-                12,
-                combine_record_and_data_type(GDSRecord::Angle, GDSDataType::EightByteReal),
-            ];
-            write_u16_array_to_file(buffer, &buffer_rot)?;
-            write_float_to_eight_byte_real_to_file(buffer, angle)?;
-        }
-    }
-
-    Ok(())
-}
 
 #[allow(clippy::too_many_lines)]
 pub fn from_gds<P: AsRef<std::path::Path>>(
