+++ crates/gdsr/src/io/write/gds_format.rs
