+++ crates/gdsr/src/io/write/svg.rs
