+++ crates/gdsr/src/lib.rs
@@ -4,6 +4,7 @@ mod elements;
 mod error;
 mod geometry;
 mod grid;
+pub(crate) mod io;
 mod library;
 mod point;
 #[cfg(test)]
@@ -13,14 +14,14 @@ mod traits;
 mod transformation;
 mod types;
 mod units;
-mod utils;
-mod writer;
 
 pub use cell::Cell;
 pub use elements::text::{HorizontalPresentation, VerticalPresentation};
 pub use elements::{Element, GdsBox, Instance, Node, Path, PathType, Polygon, Reference, Text};
 pub use error::GdsError;
 pub use grid::Grid;
+pub use io::write::svg::cell_to_svg;
+pub use io::write::{GdsFileWriter, GdsWriter};
 pub use library::{DanglingCellReference, Library};
 pub use point::Point;
 pub use stats::{CellStats, LibraryStats};
@@ -28,5 +29,3 @@ pub use traits::{Dimensions, Movable, Transformable};
 pub use transformation::{Reflection, Rotation, Scale, Transformation, Translation};
 pub use types::{AngleInRadians, DataType, Layer, LayerMapping};
 pub use units::{DEFAULT_FLOAT_UNITS, DEFAULT_INTEGER_UNITS, FloatUnit, IntegerUnit, Unit};
-pub use utils::svg::cell_to_svg;
-pub use writer::{GdsFileWriter, GdsWriter};
