--- crates/gdsr/src/utils/mod.rs
@@ -1,3 +1,0 @@
-pub mod gds_format;
-pub mod io;
-pub mod svg;
